Support passing credentials via --prefs file
Removed the input parameters for the `APPLE_ID` and `PASSWORD_FILE` variables and added the `password` variable to the `AppleDataGatherer` processor step.

This was done so that these values can be passed via autopkg --prefs file or via an override.  Currently the `password` input variable for `AppleDataGatherer` cannot be used in this recipe and overriding it does not allow for this either.